modify loss weight bug

+++ train.py
@@ -26,7 +26,7 @@ def train(gpu_id: int, param_file: str, ts_fig_dir: str, ckpt_file: Optional[str
 
     if ckpt_file is None:
         datamodule.setup("fit")
-        model = CNN3(param, torch.from_numpy(len(datamodule.dataset["train"].label) / datamodule.dataset["train"].breakdown))
+        model = CNN3(param, torch.from_numpy(len(datamodule.dataset["train"].label) / datamodule.dataset["train"].breakdown).to(dtype=torch.float32))
         trainer.fit(model, datamodule=datamodule)
         model.load_from_checkpoint(glob(path.join(trainer.log_dir, "checkpoints/", "epoch=*-step=*.ckpt"))[0], loss_weight=torch.empty(10, dtype=torch.float32))
     else:
